added new commit

+++ src/main/java/by/du/petrolstation/model/Order.java
@@ -19,7 +19,7 @@ public class Order {
     private Long id;
 
     private BigDecimal amount;
-    private Integer count;
+    private BigDecimal count;
     @OneToOne
     private Petrol petrol;
 }
